MockPeerId: Deserialize + Serialize

+++ handel/tests/mod.rs
@@ -246,7 +246,7 @@ async fn it_can_aggregate() {
     // Initialize `contributor_num networks and Handel Aggregations. Connect all the networks with each other.
     for id in 0..contributor_num {
         // Create a network with id = `id`
-        let net = Arc::new(hub.new_network_with_address(id));
+        let net = Arc::new(hub.new_network_with_address(id as u64));
         // Create a protocol with `contributor_num + 1` peers set its id to `id`. Require `contributor_num` contributions
         // meaning all contributions need to be aggregated with the additional node initialized after this for loop.
         let protocol = Protocol::new(id, contributor_num + 1, contributor_num);
@@ -288,7 +288,7 @@ async fn it_can_aggregate() {
     }
 
     // same as in the for loop, except we want to keep the handel instance and not spawn it.
-    let net = Arc::new(hub.new_network_with_address(contributor_num));
+    let net = Arc::new(hub.new_network_with_address(contributor_num as u64));
     let protocol = Protocol::new(contributor_num, contributor_num + 1, contributor_num + 1);
     let mut contributors = BitSet::new();
     contributors.insert(contributor_num);
